On this commit I in general made the code in 3dFileread prettier and put a lot of the stuff in 3dstuffLib which is meant to be my personal Library for this stuff. Also I did some general reorganization.
I am now working on 3d lighting engine over a flatplane
which is in progress and I also commited the incomplete
code for it.

Also A Vector3D lib has also been added which allows
the programmer to deal with 3d vectors with incredible
ease. And it uses some hardcore object orienting wwhich
I am quite proud of.